requirements.txt
asgiref==3.5.1
cx-Oracle==8.3.0
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.7
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
